feat(holdings): share the single-filer clone backtest via a BusinessLogic provider
Extract the on-demand single-filer clone backtest into HoldingsCloneBacktestProvider
in Equibles.Holdings.BusinessLogic so a non-web caller (an MCP tool) can run it. The
provider resolves the filer and benchmark, selects the relevant 13F snapshots, and
values them forward through the existing look-ahead-safe price loader and calculator.

HoldingsBacktestService now delegates to it and keeps only the view-model mapping and
the benchmark picker, so the portal profile page output is unchanged. The snapshot-
selection unit tests move with the methods to the provider; the forward-fill tests are
already covered by BacktestPriceLoaderForwardFill*, with the empty-series edge relocated
alongside them.
